+++ tests/test_general.py
@@ -118,10 +118,12 @@ async def test_industrial_groups_overview():
 @file('weatherforecast.json')
 async def test_market_map_data():
     df = await market_map_data()
+    if df.empty:
+        return
     assert len(df) > 300
     assert not df.lVal18AFC.str.contains('ي').any()
     assert [*df.dtypes.items()] == [
-        ('insCode', dtype('int64')),
+        ('insCode', string),
         ('dEven', dtype('int64')),
         ('hEven', dtype('int64')),
         ('pClosing', dtype('float64')),