+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 * BREAKING CHANGE: Require pyarrow and pandas >= 2.1.0 and use pyarrow string dtype in pandas. Also ``pandas.options.future.infer_string = True``.
+* BREAKING CHANGE: The ``ins_code`` column of ``general.industrial_groups_overview()`` is now of string type. (prviously it was int64)
 * Updated dataset
 * Fixed a logging bug in market_watch_plus causing exception in rare occasions.